fix(appstudioplugin): functions transplanted
